+++ app/uploaders/zip_uploader.rb
@@ -7,8 +7,19 @@ class ZipUploader < CarrierWave::Uploader::Base
 
   # Override the directory where uploaded files will be stored.
   # This is a sensible default for uploaders that are meant to be mounted:
-  def store_dir
+  # TODO: Remove in v75
+  def former_store_dir
     File.join(model.class.to_s.underscore.pluralize, model.id.to_s, mounted_as.to_s)
+  end
+  # END TODO: Remove in v75
+
+  def store_dir
+    File.join(
+      "projects",
+      model.project_id.to_s,
+      model.class.to_s.underscore.pluralize,
+      model.id.to_s
+    )
   end
 
   # Provide a default URL as a default if there hasn't been a file uploaded:
@@ -18,7 +29,7 @@ class ZipUploader < CarrierWave::Uploader::Base
 
   # Add a white list of extensions which are allowed to be uploaded.
   def extension_whitelist
-    %w(zip)
+    %w(zip xls) # TODO: Remove "xls" in v75.
   end
 
   # Override the filename of the uploaded files: