+++ app/models/validation/in_memory_grid.rb
@@ -3,14 +3,13 @@
 module Validation
   # Represents an in memory grid that mimics a Grid saved in the database.
   class InMemoryGrid
-    attr_accessor :variable, :parent_variable, :position, :value, :response_file, :responses
+    attr_accessor :variable, :parent_variable, :position, :value, :responses
 
-    def initialize(parent_variable, position, variable, value: nil, response_file: nil, responses: [])
+    def initialize(parent_variable, position, variable, value: nil, responses: [])
       @parent_variable = parent_variable
       @position = position
       @variable = variable
       @value = value
-      @response_file = response_file
       @responses = responses.collect { |r| Validation::InMemoryResponse.new(r.value) }
     end