+++ db/migrate/20190502180441_remove_response_file_from_grids.rb
@@ -1,0 +1,5 @@
+class RemoveResponseFileFromGrids < ActiveRecord::Migration[6.0]
+  def change
+    remove_column :grids, :response_file, :text
+  end
+end