+++ db/schema.rb
@@ -10,7 +10,7 @@
 #
 # It's strongly recommended that you check this file into your version control system.
 
-ActiveRecord::Schema.define(version: 2019_03_26_211254) do
+ActiveRecord::Schema.define(version: 2019_05_02_180441) do
 
   # These are extensions that must be enabled in order to support this database
   enable_extension "plpgsql"
@@ -605,7 +605,6 @@ ActiveRecord::Schema.define(version: 2019_03_26_211254) do
     t.bigint "sheet_variable_id"
     t.bigint "variable_id"
     t.text "value"
-    t.text "response_file"
     t.bigint "user_id"
     t.integer "position"
     t.datetime "created_at"