+++ test/models/export_test.rb
@@ -111,14 +111,14 @@ class ExportTest < ActiveSupport::TestCase
     sheets_with_grids = projects(:one).sheets.where(id: sheets(:has_grid))
     (_, export_file) = exports(:four).send(:generate_csv_grids, sheets_with_grids, @temp_dir, "test-export-grids.csv", true, "")
     rows = IO.readlines(export_file).collect(&:strip)
-    assert_equal '"","","","","",grid,grid,grid,grid,grid,grid,grid,grid,grid,grid,grid,grid', rows[0]
+    assert_equal '"","","","","",grid,grid,grid,grid,grid,grid,grid,grid,grid,grid,grid', rows[0]
     assert_equal "Subject,Site,Event,Design,Sheet ID,change_options,"\
-                 "var_file,var_course_work__acct101,var_course_work__econ101,"\
+                 "var_course_work__acct101,var_course_work__econ101,"\
                  "var_course_work__math123,var_course_work__phys500,var_course_work__biol327,"\
                  "height,weight,var_bmi,var_age,time_of_day", rows[1]
-    assert_equal "Code01,#{sites(:one).number_or_id},,#{designs(:has_grid).id},863765097,1,rails.png,"\
+    assert_equal "Code01,#{sites(:one).number_or_id},,#{designs(:has_grid).id},863765097,1,"\
                  ",econ101,,,,1.5,70.0,31.11,25,43199", rows[2]
-    assert_equal "Code01,#{sites(:one).number_or_id},,#{designs(:has_grid).id},863765097,2,,"\
+    assert_equal "Code01,#{sites(:one).number_or_id},,#{designs(:has_grid).id},863765097,2,"\
                  ",,,,,2.6,80.0,11.83,36,43200", rows[3]
     assert_equal 4, rows.size
   end