+++ app/uploaders/generic_uploader.rb
@@ -7,8 +7,19 @@ class GenericUploader < CarrierWave::Uploader::Base
 
   # Override the directory where uploaded files will be stored.
   # This is a sensible default for uploaders that are meant to be mounted:
-  def store_dir
+  # TODO: Remove in v75
+  def former_store_dir
     File.join(model.class.to_s.underscore.pluralize, model.id.to_s, mounted_as.to_s)
+  end
+  # END TODO: Remove in v75
+
+  def store_dir
+    File.join(
+      "projects",
+      model.project_id.to_s,
+      model.class.to_s.underscore.pluralize,
+      model.id.to_s
+    )
   end
 
   # Provide a default URL as a default if there hasn't been a file uploaded: